reafactor: userInfo(userDevice, userPage 분리예정), settings.json gitignore에 추가, userModel 하위속성 이름 변경, 파일명 모두 userInfo로 변경
