+++ .gitignore
@@ -12,4 +12,6 @@ dist
 *.tsbuildinfo
 *.yml
 
+.vscode/
+
