--- .vscode/settings.json
@@ -1,10 +1,0 @@
-{
-  "editor.formatOnSave": true,
-  "editor.formatOnSaveMode": "file",
-  "editor.defaultFormatter": "esbenp.prettier-vscode",
-  "editor.codeActionsOnSave": [
-    "source.fixAll.eslint",
-    "source.organizeImports"
-  ],
-  "eslint.validate": ["javascript", "typescript"]
-}
