+++ src/config/db.ts
@@ -1,11 +1,11 @@
 import 'dotenv/config';
 import { Sequelize } from 'sequelize';
 
-const dbPassword = process.env.DB_PASSWORD;
-const DB_NAME = 'tracker';
-const USER_NAME = 'nemo0824';
+const DB_PASSWORD = process.env.DB_PASSWORD as string;
+const DB_NAME = process.env.DB_NAME as string;
+const USER_NAME = process.env.USER_NAME as string;
 
-const sequelize = new Sequelize(DB_NAME, USER_NAME, dbPassword, {
+const sequelize = new Sequelize(DB_NAME, USER_NAME, DB_PASSWORD, {
   host: 'localhost',
   dialect: 'mysql',
   logging: false,
